# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wolfpack Auto API",
    "version": "1.0.0",
    "description": "Auto-generated from src/app/api/**/route.ts. Do not edit by hand — run `npm run openapi` to regenerate."
  },
  "servers": [
    {
      "url": "https://{dealerSlug}.wolfpackauto.com",
      "description": "Production (dealer subdomain)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "NextAuth.js session token or API key"
      }
    }
  },
  "paths": {
...
    "/api/operator/dealers": {
      "get": {
        "summary": "List dealers",
        "tags": [
          "operator"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create dealer",
        "tags": [
          "operator"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/operator/invites/accept": {
      "get": {
        "summary": "List accept",
        "tags": [
          "operator"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create accept",
        "tags": [
          "operator"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/operator/invites": {
      "get": {
        "summary": "List invites",
        "tags": [
          "operator"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create invite",
        "tags": [
          "operator"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/status": {
      "get": {
        "summary": "List status",
        "tags": [
          "status"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}